Read HTML from SD card
Change the structure of the code so that it reads the HTML for the webpage directly from the SD card rather than being stored in program memory. This would mean we could theoretically have a very intricate, detailed HTML webpage all stored in the SD card without worrying about overloading the program memory. Also looks much more clean in the Arduino code, will remove lots of lines when finished.