Grab time from client, larger HTML
Closer to re-adding all the functionality that was once in the Arduino code. Still need a way of easily sending data back and forth between the Arduino server and the phone client. Probably the best way of server to client would be adding a <script> tag in the header via the Arduino code, and then from client to server, using GETs or POSTs as was done before.